requirements.txt
absl-py==2.4.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1763409923949/work
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy==1.3.3
cycler==0.12.1
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1769745013/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h5py==3.15.1
idna==3.11
imbalanced-learn==0.14.1
ipykernel @ file:///D:/bld/bld/rattler-build_ipykernel_1770566174/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1772790189/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.3
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jupyter_client_1767954655/work
jupyter_core @ file:///D:/bld/bld/rattler-build_jupyter_core_1760643889/work
karas==0.7.0
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.46.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matlabengine==25.2.2
matplotlib==3.10.8
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.6.1
numba==0.64.0
numpy==2.4.1
opt_einsum==3.4.0
optree==0.19.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1770676243/work
pillow==12.1.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1773199142/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
protobuf==7.34.0
psutil @ file:///D:/bld/bld/rattler-build_psutil_1769678170/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pymodbus==3.11.4
pyparsing==3.3.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487136/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1771716967/work
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==80.9.0
shap==0.51.0
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sklearn-compat==0.1.5
slicer==0.0.8
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==1.5
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.10.0
tornado @ file:///D:/bld/tornado_1765836408115/work
tqdm==4.67.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1770634525285/work
websockets==16.0
Werkzeug==3.1.6
wheel==0.45.1
wrapt==2.1.1
xgboost==3.2.0
